Fix #26840 -- Move `teardown_databases` to `django.test.utils`
Moved `teardown_databases` from a method on DiscoverRunner to a utility
function.

This commit makes it easier to build custom test runners without having
to rely on DiscoverRunner, i.e.:

https://github.com/pytest-dev/pytest-django/blob/def28820dd24283ee32650f2adb8ace21d05a0a1/pytest_django/fixtures.py#L58-L59

To be consistent with other testing utilities, this patch also moves
`setup_databases` into `django.test.utils.`.
